// src/users/interface/apply-additional-info.interface.ts
export interface ApplyAdditionalInfoArgs {
  index: number;
  sex: number;
  name: string;
  company_name: string;
  company_address: string;
  longitude: number;
  latitude: number;
  company_scale: string;
  company_kind: string;
  big_company_kind: string;
  mid_company_kind: string;
  group: number;
  job: string;
  univ_kind: string;
  univ_loc: string;
  univ_name: string;
  univ_main_branch: string;
  major: string;
  univ_day_night: string;
  field: string;
  education: string;
}